requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
python-dotenv==1.0.1
requests==2.32.3